created global error handling
