+++ backend/services/todo/todo.controller.js
@@ -1,32 +1,17 @@
 import TodoItem from './todo.models.js';
 import TodoMapper from './todo.mapper.js';
-import { validationResult } from 'express-validator';
 
 const fetchAllTodos = async (_, res) => {
     const todoItems = await TodoItem.find({deletedAt: null});
     res.send({body: todoItems.map(TodoMapper.toDto)});
 };
 
-const createTodo = async (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-    }
-
-    try{
-        const todoItem = await TodoItem.create(TodoMapper.fromCreateRequest(req.body)).catch(next);
-        res.send({body: TodoMapper.toDto(todoItem)});
-    } catch(error){
-        next(error)
-    }
+const createTodo = async (req, res) => {
+    const todoItem = await TodoItem.create(TodoMapper.fromCreateRequest(req.body));
+    res.status(201).send({body: TodoMapper.toDto(todoItem)});
 };
 
 const updateTodo = async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-    }
-
     const todoItem = await TodoItem.findOneAndUpdate({id: req.params.id}, TodoMapper.fromUpdateRequest(req.body), {new: true});
     res.send({body: TodoMapper.toDto(todoItem)});
 };
