+++ backend/services/todo/todo.routes.js
@@ -1,18 +1,12 @@
 import express from 'express';
-import {body} from 'express-validator';
 import todoController from './todo.controller.js';
+import {todoValidator, errorHandler} from './todo.validator.js';
 
 const router = express.Router();
 
-const validator = [
-    body("id").exists().isString().notEmpty(),
-    body("name").exists().isString().notEmpty(),
-    body("isComplete").optional().isBoolean()
-]
-
-router.get('/', todoController.fetchAllTodos);
-router.post('/', validator, todoController.createTodo);
-router.put('/:id', validator, todoController.updateTodo);
-router.delete('/:id', todoController.deleteTodo);
+router.get('/', errorHandler(todoController.fetchAllTodos));
+router.post('/', todoValidator, errorHandler(todoController.createTodo));
+router.put('/:id', todoValidator, errorHandler(todoController.updateTodo));
+router.delete('/:id', errorHandler(todoController.deleteTodo));
 
 export default router;
